refactor: try to make every check happy
