+++ marianne/imagesearch/__init__.py
@@ -1,5 +1,2 @@
 """Top-level module for marianne/imagesearch"""
 # marianne/imagesearch/__init__.py
-
-from .descriptor import Descriptor
-from .searcher import Searcher
