+++ utils/unsplash_download.py
@@ -1,24 +1,23 @@
 import urllib.request
-from time import sleep
+from multiprocessing.pool import ThreadPool
+from pathlib import Path
 
 import pandas as pd
-
-from pathlib import Path
-from multiprocessing.pool import ThreadPool
 
 unsplash_dataset_path = Path("instance/data/unsplash-lite")
 
 # Read the photos table
-photos = pd.read_csv(unsplash_dataset_path / "photos.tsv000", sep='\t', header=0)
+photos = pd.read_csv(unsplash_dataset_path / "photos.tsv000", sep="\t", header=0)
 
 # Extract the IDs and the URLs of the photos
-photo_urls = photos[['photo_id', 'photo_image_url']].values.tolist()
+photo_urls = photos[["photo_id", "photo_image_url"]].values.tolist()
 
 # Print some statistics
-print(f'Photos in the dataset: {len(photo_urls)}')
+print(f"Photos in the dataset: {len(photo_urls)}")
 
 # Path where the photos will be downloaded
 photos_donwload_path = unsplash_dataset_path / "photos"
+
 
 # Function that downloads a single photo
 def download_photo(photo):
@@ -35,10 +34,11 @@ def download_photo(photo):
     if not photo_path.exists():
         try:
             urllib.request.urlretrieve(photo_url, photo_path)
-        except:
+        except BaseException:
             # Catch the exception if the download fails for some reason
             print(f"Cannot download {photo_url}")
             pass
+
 
 # Create the thread pool
 threads_count = 16
@@ -48,4 +48,4 @@ pool = ThreadPool(threads_count)
 pool.map(download_photo, photo_urls)
 
 # Print some statistics
-print(f'Photos downloaded: {len(photos)}')
+print(f"Photos downloaded: {len(photos)}")
